Added startup file to better manage app.js
