+++ src/App.js
@@ -1,52 +1,12 @@
 import "@baltimorecounty/dotgov-components/lib/styles/dotgov.min.css";
 import "./App.css";
 
-import { Config } from "@baltimorecounty/javascript-utilities";
 import { FilterList } from "@baltimorecounty/react-filter-list";
 import FireNewsRoomCard from "./components/FireNewsRoomCard";
 import React from "react";
+import { Run, getValue, filters } from "./Startup";
 
-const { setConfig, getValue } = Config;
-
-const testApiRoot =
-  "https://testservices.baltimorecountymd.gov/api/hub/structuredContent/news/fire";
-const prodApiRoot =
-  "https://services.baltimorecountymd.gov/hub/api/hub/structuredContent/news/fire";
-
-// HACK - the Config utiltiy does not account for beta.
-// TODO: This will need to be addressed when we get closer to launch
-const localApiRoot =
-  window.location.hostname.indexOf("beta") > -1
-    ? testApiRoot
-    : "http://localhost:54727/api/hub/structuredContent/news/fire";
-
-const configValues = {
-  local: {
-    apiRoot: localApiRoot,
-  },
-  development: {
-    apiRoot: testApiRoot,
-  },
-  staging: {
-    apiRoot: testApiRoot,
-  },
-  production: {
-    apiRoot: prodApiRoot,
-  },
-};
-
-setConfig(configValues);
-
-const filters = [
-  {
-    targetApiField: "category.value",
-    displayName: "Category",
-    options: [
-      { value: "releases", label: "News Releases" },
-      { value: "stories", label: "Stories" },
-    ],
-  },
-];
+Run();
 
 function App() {
   return (
